fix :  [ article server ]  문제점 보완
1. 탈퇴자 게시글 삭제시 redis에 저장 된 user 권한이 탈퇴자여야 함
  -> 탈퇴 권한이 아닌 경우 예외 던짐
2. 게시글 등록, 수정, 삭제시 요청 id와 토큰 id 비교 추가
3. ArticleSaveRequest에서 안쓰는 무의미한 articleId 필드 제거
4. articleService에서 특정 사용자 전체 게시글 조회 코드 오류 고침

issue key : SB2-26

+++ simpleBoard02/server-c-article/src/main/java/kim/zhyun/serverarticle/service/impl/ArticleServiceImpl.java
@@ -22,8 +22,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static kim.zhyun.serverarticle.data.message.ExceptionMessage.EXCEPTION_ARTICLE_NOT_FOUND;
-import static kim.zhyun.serverarticle.data.message.ExceptionMessage.EXCEPTION_NOT_FOUND;
+import static kim.zhyun.serverarticle.data.message.ExceptionMessage.*;
+import static kim.zhyun.serverarticle.data.type.RoleType.ROLE_WITHDRAWAL;
 import static org.springframework.data.domain.Sort.Order.desc;
 
 @RequiredArgsConstructor
@@ -56,7 +56,7 @@ public class ArticleServiceImpl implements ArticleService {
     public List<ArticleResponse> findAllByUser(long userId) {
         JwtUserDto jwtUserDto = getJwtUserDto(userId);
         
-        return articleRepository.findAll(Sort.by(desc("createdAt"))).stream()
+        return articleRepository.findAllByUserIdOrderByCreatedAtDesc(userId).stream()
                 .map(article -> ArticleResponse.from(article, jwtUserDto))
                 .toList();
     }
@@ -97,6 +97,10 @@ public class ArticleServiceImpl implements ArticleService {
             throw new ArticleException(EXCEPTION_ARTICLE_NOT_FOUND);
         
         Article article = articleContainer.get();
+        
+        if (article.getArticleId() != request.getArticleId() || article.getUserId() != request.getUserId())
+            throw new ArticleException(EXCEPTION_NOT_FOUND);
+        
         article.setTitle(request.getTitle());
         article.setContent(request.getContent());
     }
@@ -115,6 +119,11 @@ public class ArticleServiceImpl implements ArticleService {
     
     @Override
     public void deleteUserAll(long userId) {
+        Optional<JwtUserInfo> container = jwtUserInfoRepository.findById(userId);
+        
+        if (container.isEmpty() || !container.get().getGrade().equals(ROLE_WITHDRAWAL))
+            throw new MemberException(EXCEPTION_NOT_WITHDRAWAL);
+            
         List<Article> list = articleRepository.findAllByUserIdOrderByCreatedAtDesc(userId);
         articleRepository.deleteAllInBatch(list);
     }
